Start implementing a 64-bit lowering pass
Eventually we're looking to implement `wasm2js` here with `walrus`. JS
itself cannot handle 64-bit integer operations right now (or at least
not the target of `wasm2js` which is JS implementations without
WebAssembly). As a result we'll have to remove all 64-bit integer
operations from a wasm file.

This commit implements a pass which is intended to preserve the
operational semantics of the wasm module, except it won't actually use
64-bit integer operations. All 64-bit types and operations are lowered
to being implemented in terms of 32-bit.

The general idea of the pass is that all 64-bit expressions are
converted to an expression which evaluates the lower 32-bits of the
expressoin into a local and the expression itself becomes the upper
32-bits of the result. This pass *liberally* allocates temporaries and
blocks all over the place to ensure it's a pretty simplistic pass as
it's otherwise easy to get wrong. We'll almost certainly want some form
of cleanup passes before implementing `wasm2js`, targeted at removing
unnecessary locals, coalescing locals, removing unnecessary blocks, etc.

Internal and API changes made during this commit are:

* All `FunctionBuilder::*` methods are now also available on
  `LocalFunction`, except the block equivalents which I'm not sure how
  to best do just yet.

Hm that's... it!